Rename `recordstream` to `recordsource`
