+++ service/src/test/java/org/databiosphere/workspacedataservice/recordsource/JsonRecordSourceTest.java
@@ -1,4 +1,4 @@
-package org.databiosphere.workspacedataservice.recordstream;
+package org.databiosphere.workspacedataservice.recordsource;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
